move ssh tests to their own file
