+++ src/test/shared/extensions/ssh.test.ts
@@ -1,0 +1,31 @@
+/*!
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+import * as assert from 'assert'
+import { ChildProcess } from '../../../shared/utilities/childProcess'
+import { startSshAgent } from '../../../shared/extensions/ssh'
+
+describe('SSH Agent', function () {
+    it('can start the agent on windows', async function () {
+        // TODO: we should also skip this test if not running in CI
+        // Local machines probably won't have admin permissions in the spawned processes
+        if (process.platform !== 'win32') {
+            this.skip()
+        }
+
+        const runCommand = (command: string) => {
+            const args = ['-Command', command]
+            return new ChildProcess('powershell.exe', args).run({ rejectOnErrorCode: true })
+        }
+
+        const getStatus = () => {
+            return runCommand('echo (Get-Service ssh-agent).Status').then(o => o.stdout)
+        }
+
+        await runCommand('Stop-Service ssh-agent')
+        assert.strictEqual(await getStatus(), 'Stopped')
+        await startSshAgent()
+        assert.strictEqual(await getStatus(), 'Running')
+    })
+})
